added color and revision for meal. Start ColorLegendAPI

+++ v1/MealsAPI.py
@@ -12,16 +12,23 @@ meal_fields = {
     'name': fields.String,
     'category_id': fields.Integer,
     'group': fields.String,
-    'client_id': fields.String
+    'client_id': fields.String,
+    'revision': fields.Integer,
+    'color': fields.String
 }
 
 post_parser = reqparse.RequestParser()
 post_parser.add_argument('name', required='True')
 post_parser.add_argument('category_id', required='True')
 post_parser.add_argument('client_id', required='True')
+post_parser.add_argument('revision', required='True')
+post_parser.add_argument('color', required='True')
 
 get_parser = reqparse.RequestParser()
 get_parser.add_argument('cat')
+
+delete_parser = reqparse.RequestParser()
+delete_parser.add_argument('revision', required='True')
 
 
 class MealsListAPI(Resource):
@@ -36,7 +43,7 @@ class MealsListAPI(Resource):
             return meal, 200
 
         meal = Meal(name=args['name'], group=g.user.group,
-                    category_id=args['category_id'], client_id=args['client_id'])
+                    category_id=args['category_id'], client_id=args['client_id'], revision=1, color=args['color'])
         db.session.add(meal)
         db.session.commit()
         return meal, 201
@@ -58,19 +65,33 @@ class MealAPI(Resource):
     def put(self, meal_id):
         args = post_parser.parse_args()
         meal = Meal.query.get_or_404(meal_id)
+
         if meal.group != g.user.group:
             abort(403, message='Access denied')
+
+        if meal.revision != args['revision']:
+            abort(409, message='Wrong revision')
+
         meal.name = args['name']
         meal.client_id = args['client_id']
         meal.category_id = args['category_id']
+        meal.color = args['color']
+        meal.revision += 1
+
         db.session.add(meal)
         db.session.commit()
         return meal
 
     def delete(self, meal_id):
+        args = delete_parser.parse_args()
         meal = Meal.query.get_or_404(meal_id)
+
         if meal.group != g.user.group:
             abort(403, message='Access denied')
+
+        if meal.revision != args['revision']:
+            abort(409, message='Wrong revision')
+
         db.session.delete(meal)
         db.session.commit()
         return {'result': 'Meal deleted'}
